*: fix data race in the TestUpgradeVersion83
