+++ session/bootstrap_test.go
@@ -18,6 +18,7 @@ import (
 	"context"
 	"fmt"
 	"strconv"
+	"strings"
 	"testing"
 
 	"github.com/pingcap/tidb/config"
@@ -781,6 +782,51 @@ func TestUpgradeVersion74(t *testing.T) {
 			require.Equal(t, strconv.Itoa(ca.newValue), row.GetString(0))
 		}()
 	}
+}
+
+func TestUpgradeVersion75(t *testing.T) {
+	ctx := context.Background()
+
+	store, _ := createStoreAndBootstrap(t)
+	defer func() { require.NoError(t, store.Close()) }()
+
+	seV74 := createSessionAndSetID(t, store)
+	txn, err := store.Begin()
+	require.NoError(t, err)
+	m := meta.NewMeta(txn)
+	err = m.FinishBootstrap(int64(74))
+	require.NoError(t, err)
+	err = txn.Commit(context.Background())
+	require.NoError(t, err)
+	mustExec(t, seV74, "update mysql.tidb set variable_value='74' where variable_name='tidb_server_version'")
+	mustExec(t, seV74, "commit")
+	mustExec(t, seV74, "ALTER TABLE mysql.user DROP PRIMARY KEY")
+	mustExec(t, seV74, "ALTER TABLE mysql.user MODIFY COLUMN Host CHAR(64)")
+	mustExec(t, seV74, "ALTER TABLE mysql.user ADD PRIMARY KEY(Host, User)")
+	unsetStoreBootstrapped(store.UUID())
+	ver, err := getBootstrapVersion(seV74)
+	require.NoError(t, err)
+	require.Equal(t, int64(74), ver)
+	r := mustExec(t, seV74, `desc mysql.user`)
+	req := r.NewChunk(nil)
+	row := req.GetRow(0)
+	require.NoError(t, r.Next(ctx, req))
+	require.Equal(t, "host", strings.ToLower(row.GetString(0)))
+	require.Equal(t, "char(64)", strings.ToLower(row.GetString(1)))
+
+	domV75, err := BootstrapSession(store)
+	require.NoError(t, err)
+	defer domV75.Close()
+	seV75 := createSessionAndSetID(t, store)
+	ver, err = getBootstrapVersion(seV75)
+	require.NoError(t, err)
+	require.Equal(t, currentBootstrapVersion, ver)
+	r = mustExec(t, seV75, `desc mysql.user`)
+	req = r.NewChunk(nil)
+	row = req.GetRow(0)
+	require.NoError(t, r.Next(ctx, req))
+	require.Equal(t, "host", strings.ToLower(row.GetString(0)))
+	require.Equal(t, "char(255)", strings.ToLower(row.GetString(1)))
 }
 
 func TestForIssue23387(t *testing.T) {
