+++ session/upgrade_test.go
@@ -19,7 +19,6 @@ import (
 	"strings"
 	"testing"
 
-	"github.com/pingcap/tidb/meta"
 	"github.com/pingcap/tidb/session"
 	"github.com/pingcap/tidb/testkit"
 	"github.com/stretchr/testify/require"
@@ -55,51 +54,4 @@ func TestUpgradeVersion83(t *testing.T) {
 		require.Equal(t, statsHistoryTblFields[i].field, strings.ToLower(row.GetString(0)))
 		require.Equal(t, statsHistoryTblFields[i].tp, strings.ToLower(row.GetString(1)))
 	}
-}
-
-func TestUpgradeVersion75(t *testing.T) {
-	ctx := context.Background()
-
-	store, _, clean := testkit.CreateMockStoreAndDomain(t)
-	defer clean()
-
-	tk := testkit.NewTestKit(t, store)
-
-	seV74 := tk.Session()
-	txn, err := store.Begin()
-	require.NoError(t, err)
-	m := meta.NewMeta(txn)
-	err = m.FinishBootstrap(int64(74))
-	require.NoError(t, err)
-	err = txn.Commit(context.Background())
-	require.NoError(t, err)
-
-	tk.MustExec("update mysql.tidb set variable_value='74' where variable_name='tidb_server_version'")
-	tk.MustExec("commit")
-	tk.MustExec("ALTER TABLE mysql.user DROP PRIMARY KEY")
-	tk.MustExec("ALTER TABLE mysql.user MODIFY COLUMN Host CHAR(64)")
-	tk.MustExec("ALTER TABLE mysql.user ADD PRIMARY KEY(Host, User)")
-	session.UnsetStoreBootstrapped(store.UUID())
-	ver, err := session.GetBootstrapVersion(seV74)
-	require.NoError(t, err)
-	require.Equal(t, int64(74), ver)
-	r, err := tk.Exec(`desc mysql.user`)
-	require.NoError(t, err)
-	req := r.NewChunk(nil)
-	row := req.GetRow(0)
-	require.NoError(t, r.Next(ctx, req))
-	require.Equal(t, "host", strings.ToLower(row.GetString(0)))
-	require.Equal(t, "char(64)", strings.ToLower(row.GetString(1)))
-
-	tk2 := testkit.NewTestKit(t, store)
-	ver, err = session.GetBootstrapVersion(tk2.Session())
-	require.NoError(t, err)
-	require.Equal(t, session.CurrentBootstrapVersion, ver)
-	r, err = tk2.Exec(`desc mysql.user`)
-	require.NoError(t, err)
-	req = r.NewChunk(nil)
-	row = req.GetRow(0)
-	require.NoError(t, r.Next(ctx, req))
-	require.Equal(t, "host", strings.ToLower(row.GetString(0)))
-	require.Equal(t, "char(255)", strings.ToLower(row.GetString(1)))
 }
